+++ src/test/cli.up.test.ts
@@ -26,9 +26,11 @@ describe('Dev Containers CLI', function () {
 
 	describe('Command up', () => {
 		it('should execute successfully with valid config', async () => {
-			const res = await shellExec(`${cli} up --workspace-folder ${__dirname}/configs/image`);
+			const res = await shellExec(`${cli} up --workspace-folder ${__dirname}/configs/image --include-configuration --include-merged-configuration`);
 			const response = JSON.parse(res.stdout);
 			assert.equal(response.outcome, 'success');
+			assert.equal(response.configuration?.remoteEnv?.TEST_RE, 'TEST_VALUE3');
+			assert.equal(response.mergedConfiguration?.remoteEnv?.TEST_RE, 'TEST_VALUE3');
 			const containerId: string = response.containerId;
 			assert.ok(containerId, 'Container id not found.');
 			await shellExec(`docker rm -f ${containerId}`);
@@ -90,6 +92,85 @@ describe('Dev Containers CLI', function () {
 			after(async () => await devContainerDown({ composeProjectName: upResult?.composeProjectName }));
 			it('should succeed', () => {
 				assert.equal(upResult!.outcome, 'success');
+			});
+		});
+		describe('for minimal docker-compose with custom project name', () => {
+			let upResult: UpResult | null = null;
+			const testFolder = `${__dirname}/configs/compose-with-name`;
+			before(async () => {
+				// build and start the container
+				upResult = await devContainerUp(cli, testFolder, { 'logLevel': 'trace', extraArgs: `--docker-compose-path trigger-compose-v2` });
+			});
+			after(async () => await devContainerDown({ composeProjectName: upResult?.composeProjectName }));
+			it('should succeed', () => {
+				assert.equal(upResult!.outcome, 'success');
+				assert.equal(upResult!.composeProjectName, 'custom-project-name');
+			});
+		});
+		describe('for minimal docker-compose with custom project name using environment variable', () => {
+			let upResult: UpResult | null = null;
+			const testFolder = `${__dirname}/configs/compose-with-name-using-env-var`;
+			before(async () => {
+				// build and start the container
+				upResult = await devContainerUp(cli, testFolder, {
+					logLevel: 'trace',
+					extraArgs: `--docker-compose-path trigger-compose-v2`,
+					env: {
+						...process.env,
+						'CUSTOM_NAME': 'custom-name-with-env-var'
+					}
+				});
+			});
+			after(async () => await devContainerDown({ composeProjectName: upResult?.composeProjectName }));
+			it('should succeed', () => {
+				assert.equal(upResult!.outcome, 'success');
+				assert.equal(upResult!.composeProjectName, 'custom-name-with-env-var');
+			});
+		});
+		describe('for minimal docker-compose with custom project name "devcontainer" using environment variable', () => {
+			let upResult: UpResult | null = null;
+			const testFolder = `${__dirname}/configs/compose-with-name-using-env-var`;
+			before(async () => {
+				// build and start the container
+				upResult = await devContainerUp(cli, testFolder, {
+					logLevel: 'trace',
+					extraArgs: `--docker-compose-path trigger-compose-v2`,
+					env: {
+						...process.env,
+						'CUSTOM_NAME': 'devcontainer'
+					}
+				});
+			});
+			after(async () => await devContainerDown({ composeProjectName: upResult?.composeProjectName }));
+			it('should succeed', () => {
+				assert.equal(upResult!.outcome, 'success');
+				assert.equal(upResult!.composeProjectName, 'devcontainer');
+			});
+		});
+		describe('for minimal docker-compose with custom project name and custom yaml', () => {
+			let upResult: UpResult | null = null;
+			const testFolder = `${__dirname}/configs/compose-with-name-and-custom-yaml`;
+			before(async () => {
+				// build and start the container
+				upResult = await devContainerUp(cli, testFolder, { 'logLevel': 'trace', extraArgs: `--docker-compose-path trigger-compose-v2` });
+			});
+			after(async () => await devContainerDown({ composeProjectName: upResult?.composeProjectName }));
+			it('should succeed', () => {
+				assert.equal(upResult!.outcome, 'success');
+				assert.equal(upResult!.composeProjectName, 'custom-project-name-custom-yaml');
+			});
+		});
+		describe('for minimal docker-compose without custom project name', () => {
+			let upResult: UpResult | null = null;
+			const testFolder = `${__dirname}/configs/compose-without-name`;
+			before(async () => {
+				// build and start the container
+				upResult = await devContainerUp(cli, testFolder, { 'logLevel': 'trace', extraArgs: `--docker-compose-path trigger-compose-v2` });
+			});
+			after(async () => await devContainerDown({ composeProjectName: upResult?.composeProjectName }));
+			it('should succeed', () => {
+				assert.equal(upResult!.outcome, 'success');
+				assert.equal(upResult!.composeProjectName, 'compose-without-name_devcontainer');
 			});
 		});